deconstructed routes from business logic & removed unnecessary use client statements
